+++ src/statistics/statistics.service.ts
@@ -1,0 +1,102 @@
+import { Injectable } from '@nestjs/common';
+import { TeamsService } from '../teams/services/teams.service';
+import { RedisService } from '../common/redis.service';
+import { AnswersService } from '../answers/services/answers.service';
+import { TeamMembersService } from '../teams/services/teamMembers.service';
+import { TeamDocument } from '../teams/models/team.schema';
+import { AnswerDocument } from '../answers/models/answer.model';
+
+@Injectable()
+export class StatisticsService {
+  constructor(
+    private readonly teamsService: TeamsService,
+    private readonly teamMembersService: TeamMembersService,
+    private readonly answerService: AnswersService,
+    private readonly redisService: RedisService,
+  ) {}
+
+  async getTeamStats(id: string): Promise<void> {
+    const teams: TeamDocument[] = await this.teamsService.findAll();
+    const teamMemberCounts =
+      (await this.teamMembersService.getTeamMemberCounts()) as any as {
+        _id: string;
+        count: number;
+      }[];
+    const counts = teamMemberCounts.map((count) => count.count);
+
+    const teamStats = {
+      teamCount: teams.length,
+      teamMembers: {
+        mean: counts.reduce((p, c) => p + c, 0) / counts.length,
+        median: this.median(counts),
+        max: Math.max(...counts),
+      },
+    };
+
+    await this.redisService.set(id, JSON.stringify(teamStats), {
+      expireIn: 60 * 60 * 24,
+    }); // set to expire in a day
+
+    return;
+  }
+
+  async getReportStats(id: string): Promise<void> {
+    const reports: AnswerDocument[] = await this.answerService.find({});
+    const reportStats: {
+      year: string;
+      count: number;
+      months: { month: string; count: number }[];
+    }[] = [];
+    reports.forEach((report) => {
+      const date = new Date(report.createdAt);
+      const year = date.getFullYear().toString();
+      const month = date.toLocaleString('default', { month: 'long' });
+      const yearObjIndex = reportStats.findIndex(
+        (yearStat) => yearStat.year === year,
+      );
+      if (yearObjIndex === -1)
+        reportStats.push({
+          year,
+          count: 1,
+          months: [{ month, count: 1 }],
+        });
+      else {
+        const yearObj = reportStats[yearObjIndex];
+        const months = yearObj.months;
+        const monthObjIndex = months.findIndex(
+          (monthStat) => monthStat.month === month,
+        );
+        if (monthObjIndex === -1)
+          months.push({
+            month,
+            count: 1,
+          });
+        else {
+          const foundMonth = months[monthObjIndex];
+          months[monthObjIndex] = {
+            month: foundMonth.month,
+            count: foundMonth.count + 1,
+          };
+        }
+        reportStats[yearObjIndex] = {
+          year: yearObj.year,
+          count: yearObj.count + 1,
+          months,
+        };
+      }
+    });
+
+    await this.redisService.set(id, JSON.stringify(reportStats), {
+      expireIn: 60 * 60 * 24,
+    }); // set to expire in a day
+
+    return;
+  }
+
+  median(array: number[]): number {
+    if (array.length === 0) return 0;
+    array = [...array].sort((a, b) => a - b);
+    const half = Math.floor(array.length / 2);
+    return array.length % 2 ? array[half] : (array[half - 1] + array[half]) / 2;
+  }
+}
